Add more version informaton bindings in v2
